# docs/swagger.yaml
basePath: /
definitions:
  github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_api_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail'
    type: object
...
  internal_api_handler.PageMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  internal_api_handler.Paged-internal_api_handler_PaymentResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_api_handler.PaymentResponse'
        type: array
      page:
        $ref: '#/definitions/internal_api_handler.PageMeta'
    type: object
...
  internal_api_handler.PaymentResponse:
    properties:
      amount:
        type: integer
      confirmed_at:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      notes:
        type: string
      quiniela_id:
        type: integer
      reference:
        type: string
      reviewed_by:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  internal_api_handler.reviewPaymentRequest:
    properties:
      notes:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the World Cup prediction game. Manage fixtures, submit
    score forecasts, and track leaderboards.
  title: World Cup Quiniela API
  version: "1.0"
paths:
...
  /api/v1/admin/groups/{id}/payments:
    get:
      description: Returns all payment records for the given group, regardless of
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_handler.PaymentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid group ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List payments for a group
      tags:
      - admin-groups
...
  /api/v1/admin/payments:
    get:
      description: Returns a paginated list of all payment records. Supports optional
      parameters:
      - description: Filter by status (pending, confirmed, rejected, refunded)
        in: query
        name: status
        type: string
      - description: Filter by group ID
        in: query
        name: quiniela_id
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Max records per page (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.Paged-internal_api_handler_PaymentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all payments
      tags:
      - admin-payments
  /api/v1/admin/payments/{id}/reject:
    post:
      consumes:
      - application/json
      description: Marks a pending payment deposit as rejected. The rejection reason
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Rejection notes (required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.reviewPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.PaymentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Payment not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: notes are required
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reject a payment
      tags:
      - admin-payments
  /api/v1/admin/payments/{id}/validate:
    post:
      consumes:
      - application/json
      description: Confirms a pending payment deposit. Marks the payment as confirmed
      parameters:
      - description: Payment ID
        in: path
        name: id
        type: integer
      - description: Optional review notes
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_api_handler.reviewPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.PaymentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Payment not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid payment ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate a payment
      tags:
      - admin-payments
  /api/v1/admin/payments/pending:
    get:
      description: Returns all payment records with status=pending that are awaiting
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_api_handler.PaymentResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List pending payments
      tags:
      - admin-payments
...
securityDefinitions:
  BearerAuth:
    description: 'Clerk JWT token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"